% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{adjustdata}
\alias{adjustdata}
\title{Adjust Data}
\usage{
adjustdata(data)
}
\arguments{
\item{data}{Data frame to adjust}
}
\value{
Data frame with rownames as first column
}
\description{
Adjust Data
}
