% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{Driver_Score}
\alias{Driver_Score}
\title{Calculate D3 Score}
\usage{
Driver_Score(A, r)
}
\arguments{
\item{A}{Matrix A}

\item{r}{Vector r}
}
\value{
Data frame with driver scores
}
\description{
Calculate D3 Score
}
